% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{draw_panel_function_label}
\alias{draw_panel_function_label}
\title{Outsourced function which builts the plot}
\usage{
draw_panel_function_label(data, panel_scales, coord)
}
\arguments{
\item{data}{}

\item{coord}{}
}
\description{
Outsourced function which builts the plot
}
