% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakesWithR.R
\docType{package}
\name{earthquakesWithR}
\alias{earthquakesWithR}
\alias{earthquakesWithR-package}
\title{earthquakesWithR: A visualization package for world earthquake histories}
\description{
This package maps and timelines earthquakes.
}
