% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Loads and Cleans the Earth Quake data}
\usage{
eq_clean_data()
}
\value{
it returns the data frame, cleaned and with correct dates (with date class)
}
\description{
\code{eq_clean_data} loads the data, cleans the data (including the LOCATION_NAME variable)
and gives back the data frame. There is no input to this function
}
