% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates Labes for the leaflet map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{This data frame is the whole earth quake data}
}
\value{
this function gives back a vector with character strings
}
\description{
\code{eq_map} creates a string / character combining some information to display
in the pop-up
}
\examples{
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(date) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")

}
