% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans the variable "LOCATION_NAME"}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{A data frame of NOAA significant earthquake data}
}
\value{
it returns the data frame back, but with the mutated LOCATION_NAME
}
\description{
\code{eq_location_clean} strips down from the LOCATION_NAME variable the
country. The function input is the whole earth quake data, of which it then
mutates the LOCATION_NAME.
}
