% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{geom_timeline_proto_class}
\alias{geom_timeline_proto_class}
\title{Function creates the new geom (geom_timeline)
draw_panel_function is outsourced...looks nicer}
\description{
Function creates the new geom (geom_timeline)
draw_panel_function is outsourced...looks nicer
}
\keyword{datasets}
