% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{handle_negative_dates}
\alias{handle_negative_dates}
\title{Handling negative dates (BC dates)}
\usage{
handle_negative_dates(y, m, d)
}
\arguments{
\item{y}{The year as a nummeric input}

\item{m}{The month as a nummeric input}

\item{d}{the day as a numeric input}
}
\value{
Date as a date class, and which is negative
}
\description{
\code{handle_negative_dates} takes negative dates
}
\details{
This function takes ONLY negative dates in form three inputs:
year, month and the day. With some estimates and mirroring, it
gives back a date with a date-class, which is negative
Important to note is, that the return date is not 100% precise.
It is within +/- 5 days (accuracy). This is due to the mirroring
which was an estimate, and due to the fact that not all month were
equal in days.
}
\examples{
handle_negative_dates(-100, 1, 1)
}
