% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{handle_positive_dates}
\alias{handle_positive_dates}
\title{Handling positive dates (BC dates)}
\usage{
handle_positive_dates(y, m, d)
}
\arguments{
\item{y}{The year as a nummeric input}

\item{m}{The month as a nummeric input}

\item{d}{the day as a numeric input}
}
\value{
Date as a date class, and which is positive
}
\description{
\code{handle_positive_dates} takes positive dates
}
\details{
This function takes a positive date via an input of
year, month and day (seperately) and gives back the date
as a date class. very simple function.
}
\examples{
handle_positive_dates(2010, 10, 10)
}
