% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOAAearthquakes}
\alias{NOAAearthquakes}
\title{NOAA Significant Earthquakes - CLEANED version}
\format{A data frame of 5952 observations of 48 variables.  For variable
  definitions please see the "Event Variable Definitions" link at the url
  below.}
\source{
National Geophysical Data Center / World Data Service (NGDC/WDS):
  Significant Earthquake Database. National Geophysical Data Center, NOAA.
  doi:10.7289/V5TD9V7K
  \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
NOAAearthquakes
}
\description{
This is the cleaned version of the earth quakes data.
}
\keyword{datasets}
