% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Function which builts the layer for the ggplot}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{See \code{ggplot2} \code{\link[ggplot2]{layer}}}

\item{data}{See \code{ggplot2} \code{\link[ggplot2]{layer}}}

\item{stat}{See \code{ggplot2} \code{\link[ggplot2]{layer}}}

\item{position}{See \code{ggplot2} \code{\link[ggplot2]{layer}}}

\item{na.rm}{See \code{ggplot2} \code{\link[ggplot2]{layer}}}

\item{show.legend}{See \code{ggplot2} \code{\link[ggplot2]{layer}}}

\item{inherit.aes}{See \code{ggplot2} \code{\link[ggplot2]{layer}}}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}}
}
\description{
Function which builts the layer for the ggplot
}
